refactor(nodedb): split value.rs, alter.rs, and point.rs into submodules
value.rs (509 lines) → value/ with assignments, convert, defaults,
msgpack_write, rows, and time_range submodules.

alter.rs (441 lines) → alter/ with add_column, alter_type, drop_column,
enforcement, materialized_sum, and rename_column submodules. DDL router
schema.rs gains explicit routing for ADD COLUMN, DROP COLUMN, and
RENAME COLUMN so requests reach the handlers in the new module layout.

point.rs (696 lines) → point/ with apply_put, delete, get, put, and
update submodules.

All files remain under 200 lines. No behaviour changes.
