+++ nodedb/src/control/planner/sql_plan_convert/value/mod.rs
@@ -1,0 +1,21 @@
+//! Value conversion utilities: SqlValue ↔ nodedb_types::Value, msgpack encoding,
+//! column default evaluation, and WHERE-clause time-range extraction.
+
+pub(super) mod assignments;
+pub(super) mod convert;
+pub(super) mod defaults;
+pub(super) mod msgpack_write;
+pub(super) mod rows;
+pub(super) mod time_range;
+
+pub(super) use assignments::assignments_to_update_values;
+pub(super) use convert::{
+    sql_value_to_bytes, sql_value_to_msgpack, sql_value_to_nodedb_value, sql_value_to_string,
+};
+pub(super) use defaults::evaluate_default_expr;
+pub(super) use msgpack_write::{
+    row_to_msgpack, write_msgpack_array_header, write_msgpack_map_header, write_msgpack_str,
+    write_msgpack_value,
+};
+pub(super) use rows::rows_to_msgpack_array;
+pub(super) use time_range::extract_time_range;
